Improved tracking of whether a drag is active in the editor
The previous approach failed in cases where elements were added/removed from draft editor during a drag (i.e., the `onDragLeave` handler wasn't fired in cases where a previously-entered element was removed from the DOM). This happens because we dynamically update the draft editor state during the drag, which changes the DOM structure.

This new approach tracks the event targets of the drag enter/leave events. If any drag target is no longer in the document, we do not count it when determining if there is any active entered drag.