More conservative selection updating
Don't update the selection on a leaf if it only intersects the selection on the leaf's trailing edge and there are additional leaves in the block.

Before this change, in this case, we were clearing the DOM selection and adding a point to it only to immediately clear it again when the following leaf processed. In this case we do not need to update the DOM selection at all (it isn't changing) so this was 100% unnecessary overhead. It has a significant performance impact in long editor states.