Remove unnecessary DOM selection updates (#31)
* Update build

* updates jest snapshots for latest jest format

* More conservative selection updating

Don't update the selection on a leaf if it only intersects the selection on the leaf's trailing edge and there are additional leaves in the block.

Before this change, in this case, we were clearing the DOM selection and adding a point to it only to immediately clear it again when the following leaf processed. In this case we do not need to update the DOM selection at all (it isn't changing) so this was 100% unnecessary overhead. It has a significant performance impact in long editor states.

* Update package.json

* Symmetric check for more conservative selection updates

This handles the opposite case: where the selection would be set on a node that is only on the trailing edge of the selection when there are already other leaves in the block within the selection. There's no reason that this should be added to the selection.

* v0.11.6-descript.19

* v0.11.6-descript.20

* v0.11.6-descript.21

* v0.11.6-descript.22

* Option to update or set DOM selection in a single place

* more cleanup

* re-use new function to get dom selection

* v0.11.6-descript.23

* fix doc

* better docs